Share test server between packages
